backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.29.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0
google-genai>=0.2.0
sse-starlette==1.8.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
